- make len a size_t.  why should it be long ?
git-svn-id: http://unworkable.googlecode.com/svn/trunk@184 f3d975e6-390d-11de-b310-9106b1f7c239

+++ bencode.h
@@ -1,4 +1,4 @@
-/* $Id: bencode.h,v 1.17 2006-10-17 06:15:47 niallo Exp $ */
+/* $Id: bencode.h,v 1.18 2006-10-18 21:28:13 niallo Exp $ */
 /*
  * Copyright (c) 2006 Niall O'Higgins <niallo@unworkable.org>
  *
@@ -40,7 +40,7 @@ struct benc_node {
 		long long			number;
 		struct {
 			char *value;
-			long len;
+			size_t len;
 		}				string;
 		struct {
 			char *key;
